# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Build Your Christmas API
  description: |
    REST API for **Build Your Christmas** — single-brand Christmas tree e-commerce platform
    với 3 sizes (S/M/L) × 6 styles × ~25 accessories, Christmas tree editor (live quote),
    HCM-only delivery, design save/share, và admin catalog management.

    Backend is the source of truth for:
    - authentication & user identity (single-brand: chỉ roles `buyer` và `admin`)
    - catalog (Tree, Style, Accessory)
    - tree design editor (TreeDesign with live pricing)
    - shopping cart (DesignConfig-based, không phải product-based)
    - order lifecycle (HCM-only delivery, snapshot design, atomic stock reservation)
    - payment (mock, COD + online)
    - shipment tracking (HCM Delivery in-house)
    - admin management (catalog CRUD + orders + stats + users)
  version: 2.0.0
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.buildyourchristmas.vn
    description: Production (TBD)
...
paths:
...
  /api/auth/me/avatar:
    put:
      tags: [Auth]
      summary: Update user avatar
      operationId: updateAvatar
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [avatarUrl]
              properties:
                avatarUrl: { type: string, format: uri }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ============================================================
  # USERS (saved addresses)
  # ============================================================
...
  /api/cart:
    get:
      tags: [Cart]
      summary: Get user's cart
      operationId: getCart
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  cart: { $ref: '#/components/schemas/Cart' }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/CartItem' }
    delete:
      tags: [Cart]
      summary: Clear cart (alias of /api/cart/clear)
      operationId: clearCartAlias
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
  /api/cart/items:
    post:
      tags: [Cart]
      summary: Add a design to cart
      operationId: addCartItem
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                config: { $ref: '#/components/schemas/DesignConfig' }
                designId: { type: string, description: 'Alternative to config' }
                quantity: { type: integer, minimum: 1, default: 1 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  item: { $ref: '#/components/schemas/CartItem' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/cart/items/{id}:
    patch:
      tags: [Cart]
      summary: Update cart item (quantity / checked / config)
      operationId: updateCartItem
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity: { type: integer, minimum: 1 }
                checked: { type: boolean }
                config: { $ref: '#/components/schemas/DesignConfig' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  item: { $ref: '#/components/schemas/CartItem' }
    delete:
      tags: [Cart]
      summary: Remove a cart item
      operationId: deleteCartItem
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '204': { description: Deleted }
  /api/cart/clear:
    delete:
      tags: [Cart]
      summary: Clear all cart items
      operationId: clearCart
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }

  # ============================================================
  # ORDERS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
    Unauthorized:
      description: Unauthorized — token missing or invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
...
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
    User:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        email: { type: string }
        avatarUrl: { type: string }
        roles:
          type: array
          items: { type: string, enum: [buyer, admin] }
        accountStatus: { type: string, enum: [active, suspended] }
        accountStatusReason: { type: string }
        addresses:
          type: array
          items: { $ref: '#/components/schemas/Address' }
        createdAt: { type: string, format: date-time }
    Address:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        phone: { type: string }
        address: { type: string }
        province: { type: string }
        district: { type: string }
        ward: { type: string }
        isDefault: { type: boolean }
...
    DesignConfig:
      type: object
      required: [treeId, styleId, accessories, deliveryOption]
      properties:
        treeId: { type: string, description: 'Mongo ObjectId of Tree' }
        styleId: { type: string, description: 'Mongo ObjectId of Style' }
        accessories:
          type: array
          items:
            type: object
            required: [accessoryId, quantity]
            properties:
              accessoryId: { type: string, description: 'Mongo ObjectId of Accessory' }
              quantity: { type: integer, minimum: 1 }
              personalizationText: { type: string, description: 'Required if accessory.isPersonalizable. Regex ^[A-Za-z0-9 ]{1,20}$' }
        deliveryOption: { type: string, enum: [READY_TO_DISPLAY, DIY_KIT, SEPARATE] }
    PriceBreakdown:
      type: object
      properties:
        tree:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            size: { type: string, enum: [S, M, L] }
            unitPrice: { type: integer, description: 'VND' }
        style:
          type: object
          properties:
            id: { type: string }
            code: { type: string, enum: [CLASSIC, MINIMAL, GINGERBREAD, WINTER, CUTE, LUXURY] }
            name: { type: string }
        lines:
          type: array
          items:
            type: object
            properties:
              accessoryId: { type: string }
              name: { type: string }
              type: { type: string }
              group: { type: string, enum: [LIGHTS, ORNAMENT, DECOR, PERSONAL] }
              image: { type: string }
              unitPrice: { type: integer, description: 'VND' }
              quantity: { type: integer }
              lineTotal: { type: integer, description: 'VND' }
              isPersonalizable: { type: boolean }
              personalizationMaxLength: { type: integer }
              productionDays: { type: integer }
        decorationFee: { type: integer, description: 'VND — 0 unless READY_TO_DISPLAY' }
        shippingFee: { type: integer, description: 'VND — 30000' }
        unitTotal: { type: integer, description: 'VND — full price for 1 set' }
        productionDays: { type: integer }
        hasPersonalization: { type: boolean }
        hasService: { type: boolean, description: 'true if decoration fee > 0' }
        warnings:
          type: array
          items: { type: string }
    Design:
      type: object
      properties:
        _id: { type: string }
        name: { type: string }
        slug: { type: string }
        year: { type: integer }
        config: { $ref: '#/components/schemas/DesignConfig' }
        isPublic: { type: boolean }
        isPreset: { type: boolean }
        duplicatedFrom: { type: string, nullable: true }
        previewImage: { type: string }
        ownerId: { type: string, nullable: true }
        shareUrl: { type: string }
        pricing: { $ref: '#/components/schemas/PriceBreakdown' }
    Cart:
      type: object
      properties:
        _id: { type: string }
    CartItem:
      type: object
      properties:
        _id: { type: string }
        cartId: { type: string }
        designId: { type: string, nullable: true }
        quantity: { type: integer }
        priceSnapshot: { type: integer }
        currentUnitTotal: { type: integer, nullable: true }
        priceChanged: { type: boolean }
        checked: { type: boolean }
        config: { $ref: '#/components/schemas/DesignConfig' }
        design: { $ref: '#/components/schemas/Design' }
        warning: { type: string }
...